Changed switch on if

+++ src/scripts/main.js
@@ -9,63 +9,59 @@ const removeRow = document.querySelector('.remove-row');
 const removeColumn = document.querySelector('.remove-column');
 
 container.addEventListener('click', (e) => {
-  switch (true) {
-    case (e.target.matches('.append-row')):
-      const newTr = document.querySelector('tr').cloneNode(true);
+  if (e.target.matches('.append-row')) {
+    const newTr = document.querySelector('tr').cloneNode(true);
 
-      if (tbody.children.length >= 10) {
+    if (tbody.children.length === 10) {
+      e.target.disabled = true;
+    }
+
+    if (tbody.children.length === 2) {
+      removeRow.disabled = false;
+    }
+    tbody.append(newTr);
+  }
+
+  if (e.target.matches('.remove-row')) {
+    tbody.deleteRow(-1);
+
+    if (tbody.children.length === 2) {
+      e.target.disabled = true;
+    }
+
+    if (tbody.children.length === 10) {
+      appendRow.disabled = false;
+    }
+  }
+
+  if (e.target.matches('.append-column')) {
+    [...field.rows].forEach(tr => {
+      tr.innerHTML += `
+      <tr>
+        <td></td>
+      </tr>`;
+
+      if (tr.children.length === 10) {
         e.target.disabled = true;
       }
 
-      if (tbody.children.length >= 2) {
-        removeRow.disabled = false;
+      if (tr.children.length === 2) {
+        removeColumn.disabled = false;
       }
-      tbody.append(newTr);
+    });
+  }
 
-      return;
+  if (e.target.matches('.remove-column')) {
+    [...field.rows].forEach(td => {
+      td.deleteCell(-1);
 
-    case (e.target.matches('.remove-row')):
-      tbody.deleteRow(-1);
-
-      if (tbody.children.length <= 2) {
+      if (td.children.length <= 2) {
         e.target.disabled = true;
       }
 
-      if (tbody.children.length <= 10) {
-        appendRow.disabled = false;
+      if (td.children.length >= 2) {
+        appendColumn.disabled = false;
       }
-
-      return;
-
-    case (e.target.matches('.append-column')):
-      [...field.rows].forEach(tr => {
-        tr.innerHTML += `
-        <tr>
-          <td></td>
-        </tr>`;
-
-        if (tr.children.length >= 10) {
-          e.target.disabled = true;
-        }
-
-        if (tr.children.length >= 2) {
-          removeColumn.disabled = false;
-        }
-      });
-
-      return;
-
-    case (e.target.matches('.remove-column')):
-      [...field.rows].forEach(td => {
-        td.deleteCell(-1);
-
-        if (td.children.length <= 2) {
-          e.target.disabled = true;
-        }
-
-        if (td.children.length >= 2) {
-          appendColumn.disabled = false;
-        }
-      });
+    });
   }
 });
